Remove trailing spaces
This reflects the settings in `.editorconfig`: due to those, whenever
I open and edit a file in Hecke.jl, as a side effect all trailing spaces
are removed, which unnecessarily blows up my commits (or I have to manually
remove those changes again).

So, let's get rid of them in one swoop now; new ones should hopefully not
be added as long as everyone uses an editor which honors `.editorconfig`.