add a "native" type for matrices over number fields (#482)
more a proof-of-concept, the idea is to avoid the GC
by hiding the entries in the matrix and having a finalizer for
the matrix only.

rref! and ref! is working and competitive
view, hcat, vcat: OK
conversion to and from the classical type: OK

missing:
 - native add, mul, ...
 - det

all(?) generic methods should work - but without any benefit