Fix invalidations in FacElemMon constructor
The assignment `z.conj_log_cache = ...` is lowered to a call to
`setproperty!(::FacElemMon{AnticNumberField}, ::Symbol, ::Dict{RingElem, Vector{arb}})`
and this ultimately was the source of an invalidation when Singular.jl
got loaded. We now avoid this by passing the new dicts directly as
arguments to `new`.

In the `FacElemMon{S}` there also was a bug in that it used `Dict{Int,
Dict{nf_elem, arb}}`, i.e. with `arb` instead of `Vector{arb}`. The same
issue was also present in `_conjugates_arb_log`. I ended up rewriting
the latter using `get!` to make it much simpler.