add new type for Vector{fmpz} to avoid copying
In the SMat code, rows are Vector{Int} for the indices and
Vector{fmpz} for the values (at least logically)
This puts the Vector{fmpz} (julia-type) into a more
c-like array of fmpz_t*
From julia's point of view, a vector is ONE object that needs
finalizing and tracking, instead of one fmpz per entry as before.
The row-manipulation code then operates with pointers into that array
to further avoid copying of data

The SMat and SRow types are now always fully iitialized hence julia
does not use the getentry function to access structure members but just
the data offset

A SMat now can store some tmp-rows to be used in the HNF code.

plus neccessary changes in the class group...