proposal for GRH and similar - DRAFT and such
do

Hecke.set_use_GRH(true)

if you want to use GRH automatically. Possibly needs to be done using
Preferences.jl

If, e.g. by doing class_group(O) a computation is triggered using the
GRH, this is recorded

Look at

Hecke.GRH_stack

to see where.

To make use of this feature, a signature
... GRH::Bool = false, ...
needs to be changed to
... GRH::Bool = use_GRH(), ...