move code from `Oscar.jl/experimental/GModule/Misc.jl` here
The idea is to add this code to Hecke.jl,
and then to remove it from Oscar.jl.

Once we agree what shall be really added here,
tests and documentation has to be added.
(I am quite sure that I have put certain parts into the wrong files,
and that some functions have to be changed.
For example:
- The function `dual` belongs to Hecke.jl, but the method in question deals with objects from AbstractAlgebra.jl; should `dual` better get defined in AbstractAlgebra.jl?
- `pseudo_inv` is defined in AbstractAlgebra.jl, the method in question uses `MapFromFunc`, which belongs to Hecke.jl; what is the relation between Hecke.jl's `MapFromFunc` and AbstractAlgebra.jl's `FunctionalMap`?)