Fixes Serre (improves #1428)
Given an element of norm 1, Hilbert 90 assures it can be written as
sigma(x)/x
Generically, this is solves via Lagrage resolvent and randomization.
(frobenius_equation). If this fails, a much slower method is called\that
constructs the element digit-by-digit (frobenius_equation2).
The 1st function has to increase (temporarily) the precision. However
then the norm 1 condition is no longer true... and thus the input for
the second function illegal hence the failure.